+++ src/components/NumberValue.tsx
@@ -1,3 +1,5 @@
+import styled from 'styled-components';
+
 import { formatZeroNumbers } from '@/lib/formatZeroNumbers';
 
 export type NumberValueProps = {
@@ -12,16 +14,13 @@ export const NumberValue = ({ className, value, withSubscript }: NumberValueProp
 
   if (withSubscript) {
     return (
-      <span className={className}>
+      <span className={className} tw="leading-none">
         {currencySign}
         {significantDigits}
         {punctuationSymbol}
         {Boolean(zeros) && (
           <>
-            0
-            <sub title={value} tw="text-[0.85em]">
-              {zeros}
-            </sub>
+            0<$Sub title={value}>{zeros}</$Sub>
           </>
         )}
         {decimalDigits}
@@ -31,3 +30,8 @@ export const NumberValue = ({ className, value, withSubscript }: NumberValueProp
 
   return value;
 };
+
+const $Sub = styled.sub`
+  font-size: 0.85em;
+  line-height: 0;
+`;
