+++ src/pages/settings/Settings.tsx
@@ -1,4 +1,4 @@
-import { Navigate, Route, Routes, useLocation } from 'react-router-dom';
+import { Route, Routes, useLocation } from 'react-router-dom';
 
 import {
   STRING_KEYS,
@@ -16,6 +16,7 @@ import { useStringGetter } from '@/hooks/useStringGetter';
 import { ComboboxMenu } from '@/components/ComboboxMenu';
 import { PageMenu } from '@/components/PageMenu';
 import { PageMenuItemType } from '@/components/PageMenu/PageMenuItem';
+import { DisplaySettings } from '@/views/DisplaySettings';
 import { useNetworks } from '@/views/menus/useNetworks';
 
 import { useAppDispatch, useAppSelector } from '@/state/appTypes';
@@ -59,6 +60,12 @@ const SettingsPage = () => {
       href: `${AppRoute.Settings}/${MobileSettingsRoute.Preferences}`,
       label: stringGetter({ key: STRING_KEYS.PREFERENCES }),
     },
+    {
+      value: 'display-settings-nav-item',
+      type: PageMenuItemType.Navigation,
+      href: `${AppRoute.Settings}/${MobileSettingsRoute.Display}`,
+      label: stringGetter({ key: STRING_KEYS.DISPLAY_SETTINGS }),
+    },
   ];
 
   const languages = {
@@ -101,7 +108,7 @@ const SettingsPage = () => {
           path={MobileSettingsRoute.Network}
           element={<PageMenu group="network" items={[networkMenuItems]} />}
         />
-        <Route path="*" element={<Navigate to="" replace />} />
+        <Route path={MobileSettingsRoute.Display} element={<DisplaySettings tw="px-1.5 py-1" />} />
       </Routes>
     </>
   );
