+++ src/views/MarketsDropdown.tsx
@@ -26,7 +26,6 @@ import { AssetIcon } from '@/components/AssetIcon';
 import { Button } from '@/components/Button';
 import { DropdownIcon } from '@/components/DropdownIcon';
 import { IconName } from '@/components/Icon';
-import { IconButton } from '@/components/IconButton';
 import { Output, OutputType } from '@/components/Output';
 import { Popover, TriggerType } from '@/components/Popover';
 import { ColumnDef, Table } from '@/components/Table';
@@ -183,6 +182,7 @@ const MarketsDropdownContent = ({
           <$IconButton
             onClick={() => setHasSeenElectionBannerTrupmWin(true)}
             iconName={IconName.Close}
+            fullIcon
           />
         </$MarketDropdownBanner>
       );
@@ -313,7 +313,7 @@ export const MarketsDropdown = memo(
       return () => {
         window.removeEventListener('keydown', handleKeyDown);
       };
-    }, [isOpen]);
+    }, []);
 
     return (
       <$Popover
@@ -497,11 +497,6 @@ const $AssetIcon = styled(AssetIcon)<{ $uiRefreshEnabled: boolean }>`
       --asset-icon-size: 1.5em;
     `}
   `}
-`;
-
-const $IconButton = styled(IconButton)`
-  --button-backgroundColor: transparent;
-  --button-border: none;
 `;
 
 const $FlagGradient = styled.div`
